+++ catalog/app/utils/BucketPreferences/BucketPreferences.ts
@@ -9,7 +9,12 @@ import * as tagged from 'utils/taggedV2'
 import * as YAML from 'utils/yaml'
 
 export type ActionPreferences = Record<
-  'copyPackage' | 'createPackage' | 'deleteRevision' | 'openInDesktop' | 'revisePackage',
+  | 'copyPackage'
+  | 'createPackage'
+  | 'deleteRevision'
+  | 'openInDesktop'
+  | 'revisePackage'
+  | 'writeFile',
   boolean
 >
 
@@ -81,7 +86,7 @@ export interface AthenaPreferences {
 }
 
 interface UiPreferencesInput {
-  actions?: Partial<ActionPreferences>
+  actions?: Partial<ActionPreferences> | false
   athena?: AthenaPreferences
   blocks?: Partial<BlocksPreferencesInput>
   defaultSourceBucket?: DefaultSourceBucketInput
@@ -137,6 +142,7 @@ const defaultPreferences: BucketPreferences = {
       deleteRevision: false,
       openInDesktop: false,
       revisePackage: true,
+      writeFile: true,
     },
     athena: {},
     blocks: {
@@ -176,6 +182,17 @@ const bucketPreferencesValidator = makeSchemaValidator(bucketPreferencesSchema)
 function validate(data: unknown): asserts data is BucketPreferencesInput {
   const errors = bucketPreferencesValidator(data)
   if (errors.length) throw new bucketErrors.BucketPreferencesInvalid({ errors })
+}
+
+function parseActions(actions?: Partial<ActionPreferences> | false): ActionPreferences {
+  if (actions === false) {
+    return R.map(R.F, defaultPreferences.ui.actions)
+  }
+
+  return {
+    ...defaultPreferences.ui.actions,
+    ...actions,
+  }
 }
 
 function parseAthena(athena?: AthenaPreferencesInput): AthenaPreferences {
@@ -269,6 +286,7 @@ export function extendDefaults(data: BucketPreferencesInput): BucketPreferences 
   return {
     ui: {
       ...R.mergeDeepRight(defaultPreferences.ui, data?.ui || {}),
+      actions: parseActions(data?.ui?.actions),
       athena: parseAthena(data?.ui?.athena),
       blocks: parseBlocks(data?.ui?.blocks),
       packageDescription: parsePackages(